:hammer: splitting commands per file.
