+++ src/coins/rise-lisk/commands/getPubKey.h
@@ -1,0 +1,8 @@
+//
+// Created by andrea on 08/12/18.
+//
+
+#ifndef LEDGER_NANO2_GETPUBKEY_H
+#define LEDGER_NANO2_GETPUBKEY_H
+
+#endif //LEDGER_NANO2_GETPUBKEY_H
